Title: 바이러스 검사, Time: 358ms, Memory: 4MB, Status: Passed - CodeTree

+++ 240822/바이러스 검사/virus-detector.cpp
@@ -6,7 +6,7 @@ using namespace std;
 int n;
 vector<int> v;
 int x, y;
-int cnt;
+long long cnt;
 
 int main() 
 {
